inference/lambda_function/requirements.txt
tflite-runtime==2.13.0
Pillow>=10.0.0
numpy>=1.24.3
